pyproject.toml
[project]
name = "specflow"
version = "0.2.0"
description = "Spec-driven development framework for ASPICE-compliant verification"
readme = "README.md"
license = "MIT"
authors = [
    { name = "longhuiberkeley", email = "longhui@berkeley.edu" }
]
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Software Development :: Documentation",
    "Topic :: Software Development :: Testing",
    "Typing :: Typed",
]

[project.urls]
Homepage = "https://github.com/Longhuiberkeley/specflow"
Documentation = "https://github.com/Longhuiberkeley/specflow/tree/main/docs"
Repository = "https://github.com/Longhuiberkeley/specflow"
Issues = "https://github.com/Longhuiberkeley/specflow/issues"

[project.scripts]
specflow = "specflow.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]
